refactor: move git and logger classes to the services folder
