+++ src/files/json-package.ts
@@ -11,7 +11,7 @@ import {
 
 import { fileExists } from "../utils/file-state";
 import type { ForkConfig } from "../config/types";
-import type { Logger } from "../utils/logger";
+import type { Logger } from "../services/logger";
 import type { FileState, IFileManager } from "./file-manager";
 
 /** The things we are interested in, in package.json-like files. */
