+++ src/files/ms-build-project.ts
@@ -4,7 +4,7 @@ import * as cheerio from "cheerio/slim";
 
 import { fileExists } from "../utils/file-state";
 import type { ForkConfig } from "../config/types";
-import type { Logger } from "../utils/logger";
+import type { Logger } from "../services/logger";
 import type { FileState, IFileManager } from "./file-manager";
 
 /**
