+++ src/process/tag.ts
@@ -1,7 +1,7 @@
 import { formatCommitMessage } from "../utils/format-commit-message";
 import type { ForkConfig } from "../config/types";
-import type { Logger } from "../utils/logger";
-import type { Git } from "../utils/git";
+import type { Logger } from "../services/logger";
+import type { Git } from "../services/git";
 
 export async function tagChanges(
 	config: ForkConfig,
